labels: Open code strings.SplitN(,,2)
We only split into two using a single character delimiter, so it is
possible to find the delimiter with strings.IndexByte() instead of
using the more generic strings.SplitN(). SplitN() always allocates a new
slice of strings for the return value, so using IndexByte() is also
faster and more GC friendly.

Also fixes parsing of reserved labels so that, e.g., "reservedz=foo"
is not considered having a reserved source.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>